Валидация пользователя (некорректно проверяется пароль и дубликат логина)

+++ src/main/java/ru/colibri/colibriserver/controller/AdminController.java
@@ -1,23 +1,21 @@
 package ru.colibri.colibriserver.controller;
 
-import org.hibernate.Session;
-import org.hibernate.SessionFactory;
-import org.hibernate.internal.SessionFactoryImpl;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.ui.Model;
+import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
+import ru.colibri.colibriserver.domain.TestClass;
 import ru.colibri.colibriserver.security.CustomUserDetailsService;
 import ru.colibri.colibriserver.security.RoleRepository;
+import ru.colibri.colibriserver.security.UserValidator;
 import ru.colibri.colibriserver.security.UserRepository;
 import ru.colibri.colibriserver.security.model.Role;
 import ru.colibri.colibriserver.security.model.User;
 
-import javax.persistence.EntityManager;
-import javax.persistence.PersistenceContext;
 import java.util.HashSet;
 import java.util.Set;
 
@@ -29,10 +27,12 @@ import static ru.colibri.colibriserver.controller.test.UserController.encryptePa
 @Transactional
 public class AdminController {
 
-    @PersistenceContext
-    private EntityManager entityManager;
+
 
     private final Logger log = LoggerFactory.getLogger(this.getClass());
+
+    @Autowired
+    private UserValidator userValidator;
 
     @Autowired
     private CustomUserDetailsService customUserDetailsService;
@@ -43,6 +43,9 @@ public class AdminController {
     @Autowired
     private RoleRepository roleRepository;
 
+    @Autowired
+    TestClass testClass;
+
     @RequestMapping("/admin_menu")
     public String showAdminMenu(Model model) {
 
@@ -52,24 +55,32 @@ public class AdminController {
 
     @RequestMapping(value = "user_list", method = RequestMethod.GET)
     public String listUsers(Model model) {
+
+        System.err.println(testClass.getI());
+
         model.addAttribute("user", new User());
         model.addAttribute("listUsers", this.userRepository.findAllByOrderById());
-
         return "user_list";
     }
+
 
     @RequestMapping(value = "remove_user/{id}", method = RequestMethod.GET)
     public String deleteUser(@PathVariable("id") int id,
                              @RequestParam(value = "user_name", required = false) String user_name,
                              Model model) {
 
-        log.debug("Remove user: " + userRepository.findById(id));
-        userRepository.removeById(id);
+        User user = userRepository.findById(id);
 
-      //  model.addAttribute("user", new User());
+        if (user != null){
+            log.debug("Remove user: " + user);
+            userRepository.removeById(id);
+        } else {
+            user_name = null;
+        }
+
+        //  model.addAttribute("user", new User());
         model.addAttribute("listUsers", this.userRepository.findAllByOrderById());
         model.addAttribute("successRemove", user_name);
-
 
 
         return "user_list";
@@ -78,6 +89,8 @@ public class AdminController {
     @Transactional
     @RequestMapping("user_edit/{id}")
     public String userData(@PathVariable("id") int id, Model model) {
+
+
 
         User user;
 
@@ -103,16 +116,29 @@ public class AdminController {
     }
 
 
+
+
     @RequestMapping(value = "/user_edit/add", method = RequestMethod.POST)
-    public String addUser(@ModelAttribute("user") User user) {
+    public String addUser(@ModelAttribute("user") User user, Model model, BindingResult result) {
+
+        //Валидация полученных данных
+        userValidator.validate(user, result);
+        if (result.hasErrors()) {
+            log.debug("Error field of User: " + result.toString());
+            Set<Role> roles = roleRepository.findAllByOrderById();
+            model.addAttribute("userRoles", roles);
+            model.addAttribute("user", user);
+            //возваращаемся обратно с ошибкой
+            return "user_edit";
+        }
 
 
         if (user.getId() == null) {
             user.setPassword(encryptePassword(user.getPassword()));
             log.debug("Add new user: " + user.toString());
             Set<Role> roles = new HashSet<>();
-            for (Role r: user.getRoles()){
-               roles.add(roleRepository.findByRole(r.getRole()));
+            for (Role r : user.getRoles()) {
+                roles.add(roleRepository.findByRole(r.getRole()));
             }
             user.setRoles(roles);
             customUserDetailsService.saveUser(user);
